+++ tools/vulture-whitelist.py
@@ -1,2 +1,2 @@
-_.created  # unused attribute (docs/generate_log_svg.py:36)
+_.created  # unused attribute (docs/generate_log_svg.py:37)
 upload_time_iso_8601  # unused variable (src/update_time/sources/pypi.py:44)